+++ src/test/resources/benchmarks/sum_skip_stuck.rs
@@ -1,7 +1,7 @@
 // cora: MAYBE
-fn sum_skip_stuck(n: i32) -> i32 {
-    let mut i: i32 = 0;
-    let mut acc: i32 = 0;
+fn sum_skip_stuck(n: i16) -> i16 {
+    let mut i: i16 = 0;
+    let mut acc: i16 = 0;
     while i < n {
         if i == 5 {
             continue;
